Added weekly and current budget to items enum. Added reset of budget each week. Controlling bduget now possible. Decrementation of budget through shopping lists now a fact.

+++ app/controllers/shopping_lists_controller.rb
@@ -1,9 +1,9 @@
 class ShoppingListsController < ApplicationController
 	skip_before_action :verify_authenticity_token, only: [:add_item, :buy_item]  
+	require 'date'
 
 	def index
 		@family = Family.find(params[:family_id])
-		puts()
 		@list = Family.find(params[:family_id]).shopping_lists.find(params[:id])
 		@items = @list.items.where(bought:false)
 	end
@@ -42,17 +42,19 @@ class ShoppingListsController < ApplicationController
 		list = family.shopping_lists.find(params[:sl_id])
 		item = list.items.find(params[:item_id])
 		item_name = item.items_enum.name
-		
-		quantity = params[:quantity].to_i
+		category_name =  item.items_enum.category_name
+		params_price = params[:price].to_f
+		params_quantity = params[:quantity].to_i
+
 		existingItem = Item.find_by items_enum: item.items_enum, quantity_unit: item.quantity_unit, bought: true
-		
-		if !quantity
-			quantity = item.quantity
+		if params_quantity == 0
+			params_quantity = item.quantity
 		end
+
 		if not existingItem
-			if quantity < item.quantity
-				newItemQuantity = item.quantity - quantity
-				item.quantity = quantity
+			if params_quantity < item.quantity
+				newItemQuantity = item.quantity - params_quantity
+				item.quantity = params_quantity
 
 				newItem = Item.new(items_enum: item.items_enum, shopping_list: list, family: family, price: params[:price].to_i, quantity: newItemQuantity, quantity_unit: item.quantity_unit)
 				newItem.save				
@@ -60,35 +62,30 @@ class ShoppingListsController < ApplicationController
 
 			item.bought = true
 			item.save
-
-			puts item
-
-			transaction = Transaction.new
-
-			transaction.category_name = item.items_enum.category_name
-			transaction.price = params[:price].to_f * params[:quantity].to_i
-			transaction.price_unit = params[:price_unit]
-			transaction.family_id = family.id
-			
-			transaction.save
-
-			puts 'HERE'
-
-			puts transaction
-
-			puts family
-
-			family.budget -= transaction.price
-
-			family.save
-
 		else
-			existingItem.quantity += item.quantity
+			if params_quantity < item.quantity
+				quantityToAdd = params_quantity
+				item.quantity -= params_quantity
+				item.save
+			else
+				quantityToAdd =  item.quantity
+				item.destroy
+			end
+			existingItem.quantity += quantityToAdd
 			existingItem.save
-			item.destroy
 		end
-
-		redirect_to shopping_list_path(family.id, list.id), notice: "Successfully added item " + item_name + " to " + family.name + "'s current stock"
+	
+		transaction = Transaction.new
+		transaction.category_name = category_name
+		transaction.price = params_price * params_quantity
+		transaction.price_unit = "levs"
+		transaction.family_id = family.id
+		transaction.save
+		
+		family.current_budget_state -= transaction.price
+		family.save
+		
+		redirect_to shopping_list_path(family.id, list.id), notice: "Successfully added item to family's current stock"
 	end
 
 	def edit
@@ -120,6 +117,8 @@ class ShoppingListsController < ApplicationController
 
 		redirect_to "/families/" + String(@family.id) + "/shopping_lists/" + String(@list.id)
 	end
+
+private
 
 	def sl_params
 		return params.require(:shopping_list).permit(:name)
